Support fractional traffic quotas and human-readable rate limits

+++ services/api/internal/services/customer/types.go
@@ -19,7 +19,7 @@ type ModelCustomer struct {
 	ExpireAt             *time.Time `json:"expire_at" gorm:"type:date" validate:"required"`
 	DeactivatedAt        *time.Time `json:"deactivated_at" gorm:"type:date" validate:"required"`
 	TrafficType          string     `json:"traffic_type" gorm:"type:varchar(32);not null;default:1" enums:"Free,MonthlyTransmit,MonthlyReceive,MonthlyRxTx,TotallyTransmit,TotallyReceive,TotallyRxTx" validate:"required"`
-	TrafficSize          int        `json:"traffic_size" gorm:"not null" validate:"required"` // in GiB  >> x * 1024 ** 3
+	TrafficSize          int64      `json:"traffic_size" gorm:"not null" validate:"required"` // in GiB  >> x * 1024 ** 3
 	Rx                   int        `json:"rx" gorm:"not null;default:0" validate:"required"` // Receive in bytes
 	Tx                   int        `json:"tx" gorm:"not null;default:0" validate:"required"` // Transmit in bytes
 }
