+++ services/api/internal/services/ocserv_user/types.go
@@ -14,7 +14,7 @@ type CreateOcservUserData struct {
 	ExpireAt    string                   `json:"expire_at" validate:"omitempty" example:"2025-12-31"`
 	Unlimited   bool                     `json:"unlimited" validate:"omitempty" example:"false" default:"false"`
 	TrafficType string                   `json:"traffic_type" validate:"required,oneof=Free MonthlyTransmit MonthlyReceive MonthlyRxTx TotallyTransmit TotallyReceive TotallyRxTx"`
-	TrafficSize int                      `json:"traffic_size" validate:"omitempty,gte=0" example:"10737418240"` // 10 GiB
+	TrafficSize int64                    `json:"traffic_size" validate:"omitempty,gte=0" example:"10737418240"` // 10 GiB
 	Description string                   `json:"description" validate:"omitempty,max=1024" example:"User for testing VPN access"`
 	Config      *models.OcservUserConfig `json:"config" validate:"required"`
 }
@@ -25,7 +25,7 @@ type UpdateOcservUserData struct {
 	ExpireAt    *string                  `json:"expire_at"  validate:"omitempty" example:"2025-12-31"`
 	Unlimited   bool                     `json:"unlimited" validate:"omitempty" example:"false" default:"false"`
 	TrafficType *string                  `json:"traffic_type" validate:"oneof=Free MonthlyTransmit MonthlyReceive MonthlyRxTx TotallyTransmit TotallyReceive TotallyRxTx"`
-	TrafficSize *int                     `json:"traffic_size" validate:"gte=0" example:"10737418240"` // 10 GiB
+	TrafficSize *int64                   `json:"traffic_size" validate:"gte=0" example:"10737418240"` // 10 GiB
 	Description *string                  `json:"description" validate:"omitempty,max=1024" example:"User for testing VPN access"`
 	Config      *models.OcservUserConfig `json:"config" validate:"omitempty"`
 }
@@ -39,7 +39,7 @@ type SyncOcpasswdRequest struct {
 	Users       []user.Ocpasswd          `json:"users" validate:"required"`
 	ExpireAt    *string                  `json:"expire_at" validate:"omitempty" example:"2025-12-31"`
 	TrafficType *string                  `json:"traffic_type" validate:"oneof=Free MonthlyTransmit MonthlyReceive MonthlyRxTx TotallyTransmit TotallyReceive TotallyRxTx"`
-	TrafficSize *int                     `json:"traffic_size" validate:"gte=0" example:"10737418240"` // 10 GiB
+	TrafficSize *int64                   `json:"traffic_size" validate:"gte=0" example:"10737418240"` // 10 GiB
 	Description *string                  `json:"description" validate:"omitempty,max=1024" example:"User for testing VPN access"`
 	Config      *models.OcservUserConfig `json:"config" validate:"omitempty"`
 }
